FIX Change `LANG` and `LC_ALL` to `en_US.UTF-8`
Fixes #130 where we set `LANG` and `LC_ALL` to `C.UTF-8`. This works for Ubuntu images, but fails on CentOS 7 images as it does not have the locale installed. It is not easy to install on CentOS 7 (pkgs exist for CentOS 8). Making this change brings us inline with the containers used by conda-forge which also use `en_US.UTF-8`