FIX Copy .condarc instead of symlink
Despite testing in #137 that passed, when trying to build containers after the PR merge no `rapidsai` containers would build due to symlink errors. This approach copies the file instead which will result in a more consistent state if the `.condarc` file is modified for the `rapids` env, it will leave the `base` env intact.