ENH Add retry for failed gpuci-tools; drop source activate
The gpuci-tools install has no retry and even in testing this PR there are conda JSON/HTTP failures. Adding the `||` with a retry on failure will try a second time to prevent the need to rebuild the container on a temporary error.

In addition this removes all `source activate` mentions which are the old format and recent changes to `miniconda-cuda` has setup the environment to run without needing to invoke `source activate` on each `RUN` command.