Remove `BUILD_STACK_VER` from Images
This PR removes all instances of `BUILD_STACK_VER`, which is responsible for determining the `libgcc-ng` and `libstdcxx-ng` versions that are used in our images. There is already a version variable for this in the `integration` repo which is responsible for controlling the versions of these packages in `rapids-build-env`, so having them here just means unnecessary duplication.

This issue was discovered after https://github.com/rapidsai/integration/pull/410 was merged and we noticed that these images were not picking up the latest `rapids-build-env` package.