use `mamba` for environment creation & update
The move to `mamba` requires removing the `--freeze-installed` flag. When the flag was enabled, `mamba` wasn't solving the environment at all. When the flag was removed, `mamba` solved the environment, but changed a few dependency versions. However the new dependency versions were still within the version specifications that we initially set.