pin `openssl`
Conda-forge is currently migrating from OpenSSL 1.1.1 to OpenSSL 3. As part of this migration packages are being built for both versions. However not all packages have made it to OpenSSL 3 (particularly with our current version constraints around things like `nodejs`). To avoid major solve changes later in the image build, we pin to OpenSSL 1.1.1 early on to ensure minimal changes later in the environment and things like --freeze-installed behave as expected (and without lengthy solves or conflicts)
https://github.com/conda-forge/conda-forge-pinning-feedstock/pull/1896