Update `.condarc` channel priority
This PR updates the channel priority in our `.condarc` file. This is to ensure that the `cuda-python` package from `conda-forge` is preferred over the `nvidia` variant. It's important that we use the `conda-forge` variant since that depends on the `cudatoolkit` package, as opposed to the unsupported `cuda-toolkit` package which is depended on by the `nvidia` variant.